DI: Inject weather data at app launch.
